\name{naics2017}
\alias{naics2017}
\docType{data}

\title{
North American Industry Classification System (NAICS)
}

\description{
North American Industry Classification System (NAICS) Canada 2017 Version 3.0
}

\usage{data("naics2017")}

\format{
  A data frame with 2,084 observations on the following 3 variables.
  \describe{
    \item{\code{level}}{an integer vector}
    \item{\code{code}}{a character vector}
    \item{\code{class_title}}{a character vector}
  }
}

\source{
Statistics Canada, North American Industry Classification System (NAICS) Canada 2017 Version 3.0, September 14, 2018. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
}

