\name{napcs2017}
\alias{napcs2017}
\docType{data}

\title{
North American Product Classification System (NAPCS)
}

\description{
North American Product Classification System (NAPCS) Canada 2017 Version 2.0
}

\usage{data("napcs2017")}

\format{
  A data frame with 5,033 observations on the following 3 variables.
  \describe{
    \item{\code{level}}{an integer vector}
    \item{\code{code}}{a character vector}
    \item{\code{class_title}}{a character vector}
  }
}

\source{
Statistics Canada, North American Product Classification System (NAPCS) Canada 2017 Version 2.0, April 12, 2019. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
}