\docType{package}
\name{pias-package}
\alias{pias-package}
\alias{pias}

\title{Price Index Aggregation Structure}

\description{
A small package for building a price index aggregation structure from a hierarchical classification like NAICS or NAPCS. 
}

\details{
This package has one function at the moment: \code{\link[pias:pias_from_classification]{pias_from_classification()}}. This takes a hierarchical classification like NAICS and NAPSC and turns it into a list that can be used to easily aggregate a price index. The NAICS 2017 and NAPCS 2017 classifications are also packaged.
}

\seealso{
\url{https://github.com/ppd-dpp/pias}
}

\author{
\strong{Maintainer}: Steve Martin \email{steve.martin5@canada.ca}
}
